Bumped the govuk_frontend_toolkit version
This is the version that actually contains the
show/hide javascript.  It's a bit pointless updating
this because our toolkit doesn't actually use this
(the frontend apps are responsible for initiating
the javascript), but I've just updated it here for
completeness.

I thought about doing an example of this being used,
but it seems like it doesn't really work the way
oour other examples do (it includes two different
types of input fields), and it's also not really a
feature that we've built.  Instead, the example to
check out should be the one in the govuk frontend
toolkit.